Edit: add type cast to avoid warning

+++ sources/CMLStarRatingValue.m
@@ -25,7 +25,7 @@
 
 -(NSString *)debugDescription
 {
-    return [NSString stringWithFormat:@"%@: full star %ld, half star %d", [super debugDescription], self.wholeStarCount, self.hasHalfStar];
+    return [NSString stringWithFormat:@"%@: full star %ld, half star %d", [super debugDescription], (long)self.wholeStarCount, self.hasHalfStar];
 }
 
 @end
